refactor(goal-detail): split the inv-row builders out of goalDetailShared (#467)
Third slice of the goalDetailShared breakup. Move the row-building logic —
buildCompositionSegments, buildRenewalSummary, and buildInvRows (dedup + valuation
of a goal's holdings) — into a new `goalDetailRows.ts`. The data types (InvRow,
GoalDetailTx, CompositionSeg, InvTranche, RenewalSummary) STAY in goalDetailShared
since InvRow alone has ~13 importers across the app; the builders import them as
types (plus GD_COLORS as the one value dependency). This keeps the ripple to the
four builder consumers — GoalDetailSheet, DesktopGoalDetail, and two tests — which
now import the builders from goalDetailRows. The now-unused @/lib/finance +
@/lib/accumulating imports drop out of goalDetailShared. No behavior change.

Verified: full unit suite 1261 passed (goalComposition + goalDetailShared tests
exercise the builders from the new module); typecheck 0 errors; lint 0 errors;
git diff --check clean. goalDetailShared 592 → 425 lines (717 → 425 across the
breakup so far).
