+++ app/assets/components/__tests__/goalComposition.test.ts
@@ -1,5 +1,6 @@
 import { describe, it, expect } from 'vitest'
-import { buildCompositionSegments } from '../goalDetailShared'
+import {  } from '../goalDetailShared'
+import { buildCompositionSegments } from '../goalDetailRows'
 
 // The goal-detail "Cơ cấu / Composition" segments are built from the active
 // investment rows. Held-for-merge cash CLOSED its source deposit, so it is absent
